feat: handle an edge case related to `{dirname}`
If you have an alternate mapping from, say:

    "*.ts"

to:

    "{dirname}/__tests__/{basename}-test.ts"

Then the "Create alternate file" menu will show something like this:

    Create alternate file?
    1 ...
    2 ...
    3 /__tests__/example.test.ts
    4 ...
    Type number and <Enter> or click with the mouse (q or empty cancels):

After this commit, it instead shows this:

    Create alternate file?
    1 ...
    2 ...
    3 __tests__/example.test.ts
    4 ...
    Type number and <Enter> or click with the mouse (q or empty cancels):

What was going on here? Well, a pattern like "*.ts" matching a file
"example.ts" was getting decomposed as follows:

- "{dirname}" was expanding to an empty string.
- "{basename}" was "example.ts"

When we concatenate these together with "/__tests__/" in between, we got
"/__tests__/example.test.ts", which looks like an absolute path but is
not supposed to be one.

The fix comes in two parts:

1. Instead of expanding "{dirname}" to "", we expand it to "." to
   prevent it from being empty. This in itself is enough to make the
   "Create alternate file" display a relative path if we have an
   expansion like "{dirname}/{basename}.test.ts" (although you could
   write this as "{}.test.ts" and avoid the problem).
2. If our alternate starts with "./", we strip it off. This is what
   handles the specific example mentioned above.
   "./__tests__/foo.test.ts" becomes "__tests__/foo.test.ts".